+++ eVrticSistem/Controllers/OdgajateljController.cs
@@ -57,17 +57,11 @@ namespace eVrticSistem.Controllers
         // GET: Odgajatelj/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var odgajatelj = await _context.Odgajatelji
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (odgajatelj == null)
-            {
-                return NotFound();
-            }
+            if (odgajatelj == null) return NotFound();
 
             return View(odgajatelj);
         }
@@ -79,8 +73,6 @@ namespace eVrticSistem.Controllers
         }
 
         // POST: Odgajatelj/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,ImePrezime,Email,Uloga,StatusNaloga")] Odgajatelj odgajatelj)
@@ -97,30 +89,19 @@ namespace eVrticSistem.Controllers
         // GET: Odgajatelj/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var odgajatelj = await _context.Odgajatelji.FindAsync(id);
-            if (odgajatelj == null)
-            {
-                return NotFound();
-            }
+            if (odgajatelj == null) return NotFound();
             return View(odgajatelj);
         }
 
         // POST: Odgajatelj/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,ImePrezime,Email,Uloga,StatusNaloga")] Odgajatelj odgajatelj)
         {
-            if (id != odgajatelj.Id)
-            {
-                return NotFound();
-            }
+            if (id != odgajatelj.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -131,14 +112,8 @@ namespace eVrticSistem.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!OdgajateljExists(odgajatelj.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!OdgajateljExists(odgajatelj.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -148,17 +123,11 @@ namespace eVrticSistem.Controllers
         // GET: Odgajatelj/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var odgajatelj = await _context.Odgajatelji
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (odgajatelj == null)
-            {
-                return NotFound();
-            }
+            if (odgajatelj == null) return NotFound();
 
             return View(odgajatelj);
         }
@@ -169,10 +138,7 @@ namespace eVrticSistem.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var odgajatelj = await _context.Odgajatelji.FindAsync(id);
-            if (odgajatelj != null)
-            {
-                _context.Odgajatelji.Remove(odgajatelj);
-            }
+            if (odgajatelj != null) _context.Odgajatelji.Remove(odgajatelj);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
